frontend/requirements.txt
tk
watchdog
matplotlib
opencv-python
pillow
